# apps/packages/api-spec/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tag-catalog": {
      "get": {
        "operationId": "TagCatalogController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read the global tag catalog (authenticated)",
        "tags": [
          "tag-catalog"
        ]
      },
      "patch": {
        "operationId": "TagCatalogController_patchSchema",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagCatalogPatchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ops applied; returns ignored ops"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden — admin required"
          },
          "409": {
            "description": "Duplicate slug on add or rename"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Batched add/update/rename/delete of tag catalog (admin)",
        "tags": [
          "tag-catalog"
        ]
      }
    },
...
  },
  "info": {
    "title": "M.A.G.N.U.S. API",
    "description": "Server-side API for the M.A.G.N.U.S. Project",
    "version": "1.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TagCatalogEntryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "TagCatalogOpDto": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "add",
              "update",
              "rename",
              "delete"
            ]
          },
          "slug": {
            "type": "string"
          },
          "rename": {
            "type": "string",
            "description": "New slug (rename only)"
          },
          "entry": {
            "$ref": "#/components/schemas/TagCatalogEntryDto"
          }
        },
        "required": [
          "action",
          "slug"
        ]
      },
      "TagCatalogPatchDto": {
        "type": "object",
        "properties": {
          "ops": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagCatalogOpDto"
            }
          }
        },
        "required": [
          "ops"
        ]
      }
    }
  }
}